Move session shims to it's own file
