+++ src/session/sessionShims.js
